Initial version.
Add a way to define type specifications with another syntax.

Reasons for this is the more coupled way it ties to the function.
Some what inspired by flow.js but more to do with my laziness to
write @spec before the function without the need to type the function
specification twice.

It looks something like this:

```
defmodule MyMath do
  def square(n: integer) :: integer, do: n * n
  def add(a: integer, b: integer) :: integer, do: a + b
end
```

Stuff that are not supported/tested yet(TODO):
  - Multiple clause definitions
  - Pattern matching
  - Default arguments
  - Function heads